+++ libs/adapter/index.ts
@@ -1,4 +1,4 @@
-export * from './cheerio/adapter.cheerio-service';
+export * from './cheerio/cheerio-service';
 
 export * from './entity-repository/in-memory.entity-repository';
 export * from './entity-repository/mongo-db.entity-repository';
